generate "nearby" loop settings
By nearby, we actually encode some knowledge
of what the setting actually means into the generation
of the new setting.

For example, if the current setting is "disable unrolling"
for some loop, then "full unrolling" is further away than
partial unrolling.

This is achieved by mapping each discrete setting to
a range of int values [min, max]. Thus, the L1 distance
over integers indicates distance in each loop setting.

When we sample from a normal distribution centered on
one side of the range, the std deviation of the
distribution enforces the idea that a setting is
"very far".