remove error-prone optional cells for dispatching
it's a bit annoying to pass more than 1 value to
a dispatcher task. I've just gone ahead and
made structs that are locally allocated, instead
of essentially using a global std::optional,
which I think is the source of a rare bug where
std::optional raises a bad value exception.