tweaking qsort bench in a few ways
(1) It's still rather hard to tell how many
times the benchmark function is called,
and how many real iterations of the test
loop actually occur... to ensure the
Driver maintains its state while tuning.

My workaround for this confusion is to
add an unused param to the qsort function,
and fix it in the tuner so it will spawn a
new optimizer when the size of the array
changes.

(2) The experimentation rate-limiter added in
a recent commit was effectively turned off
for now, because it seems that too much
experimentation was _not_ in fact the issue,
it was some weird behavior described in (1).