a "fast" experiment is to recheck history of experiments for best
previously, the "best seen" version was defined as being the
best version only after it's (potentially short) trial phase
was done. Thus, if more measurements prove that it's actually
worse, or has become worse, we had no recourse.

Now we save the history of versions trialed, and when we
want to do a "fast" experiment, we check the history to see
if the current best is not good enough anymore, and swap
it out.

we want to avoid swapping too often, so there's a control
on that saying how much better the older one has to be
to swap it out, since cache locality is a penalty of swapping.